refactor: component & column locations
