+++ src/app/globals.css
@@ -1,6 +1,7 @@
 @import 'tailwindcss';
 @import 'tw-animate-css';
 @import 'shadcn/tailwind.css';
+@import '@stereopt/data-table/tailwind-source.css';
 
 @custom-variant dark (&:is(.dark *));
