+++ src/components/transactions/ImportTransactions.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import { EditableDataTable } from "@/components/dataTable/EditableDataTable";
-import { editableTransactionColumns } from "@/components/dataTable/columns/editableTransactionColumns";
+import { EditableDataTable } from "@/components/editableDataTable/EditableDataTable";
 import { Button } from "@/components/ui/button";
 import {
   Dropzone,
   DropzoneContent,
   DropzoneEmptyState,
 } from "@/components/ui/shadcn-io/dropzone";
+import { editableTransactionColumns } from "@/constants/columns/editableTransactionColumns";
 import { ROUTES } from "@/constants/routes";
 import { useImportTransaction } from "@/hooks/transactions/useImportTransaction";
 import { useParseTransaction } from "@/hooks/transactions/useParseTransaction";
