+++ src/components/transactions/UserTransactions.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { transactionColumns } from "@/components/dataTable/columns/transactionColumns";
 import { ErrorAlert } from "@/components/ErrorAlert";
 import { AddTransaction } from "@/components/transactions/AddTransaction";
+import { transactionColumns } from "@/constants/columns/transactionColumns";
 import { useGetTransactions } from "@/hooks/transactions/useGetTransactions";
 import { DataTable } from "@stereopt/data-table";
 import { ArrowLeftRight } from "lucide-react";
